+++ datafusion/functions-array/Cargo.toml
@@ -49,6 +49,7 @@ datafusion-functions = { workspace = true }
 itertools = { version = "0.12", features = ["use_std"] }
 log = { workspace = true }
 paste = "1.0.14"
+once_cell = "1.19.0"
 
 [dev-dependencies]
 criterion = { version = "0.5", features = ["async_tokio"] }
