+++ datafusion/physical-expr-common/src/aggregate/utils.rs
@@ -21,9 +21,10 @@ use arrow::{
     compute::SortOptions,
     datatypes::{DataType, Field},
 };
-use datafusion_physical_expr_common::sort_expr::PhysicalSortExpr;
 
-use crate::AggregateExpr;
+use crate::sort_expr::PhysicalSortExpr;
+
+use super::AggregateExpr;
 
 /// Downcast a `Box<dyn AggregateExpr>` or `Arc<dyn AggregateExpr>`
 /// and return the inner trait object as [`Any`] so
