+++ datafusion/physical-plan/Cargo.toml
@@ -47,6 +47,7 @@ datafusion-common-runtime = { workspace = true, default-features = true }
 datafusion-execution = { workspace = true }
 datafusion-expr = { workspace = true }
 datafusion-functions-aggregate = { workspace = true }
+datafusion-physical-expr-common = { workspace = true }
 datafusion-physical-expr = { workspace = true, default-features = true }
 futures = { workspace = true }
 half = { workspace = true }
