+++ datafusion/physical-plan/src/lib.rs
@@ -91,8 +91,9 @@ pub use datafusion_physical_expr::{
 pub use crate::stream::EmptyRecordBatchStream;
 pub use datafusion_execution::{RecordBatchStream, SendableRecordBatchStream};
 pub mod udaf {
-    pub use datafusion_functions_aggregate::create_aggregate_expr;
-    pub use datafusion_functions_aggregate::AggregateFunctionExpr;
+    pub use datafusion_physical_expr_common::aggregate::{
+        create_aggregate_expr, AggregateFunctionExpr,
+    };
 }
 
 /// Represent nodes in the DataFusion Physical Plan.
